Fixed str/num Arrays and some label issues need to see more

+++ ph4,5/phase4,5/src/lib/definitions.h
@@ -64,7 +64,7 @@ typedef struct SymbolTableEntry{
 	
 	union {
 		int intNum;
-		float realNum;
+		double realNum;
 		char* string;
 		short int boolean;
 		short int nil;
